+++ hyperdbg/hprdbgctrl/code/debugger/commands/meta-commands/start.cpp
@@ -57,8 +57,8 @@ CommandStart(vector<string> SplittedCommand, string Command)
         ShowMessages("the user-mode debugger in VMI Mode is still in the beta version and not stable. "
                      "we decided to exclude it from this release and release it in future versions. "
                      "if you want to test the user-mode debugger in VMI Mode, you should build "
-                     "HyperDbg with special instructions. But starting processes\n"
-                     "is fully supported in the Debugger Mode.\n"
+                     "HyperDbg with special instructions. But starting processes is fully supported "
+                     "in the Debugger Mode.\n"
                      "(it's not recommended to use it in VMI Mode yet!)\n");
         return;
     }