+++ hyperdbg/hprdbgkd/code/driver/Ioctl.c
@@ -165,27 +165,29 @@ DrvDispatchIoControl(PDEVICE_OBJECT DeviceObject, PIRP Irp)
 
             DebuggerReadMemRequest = (PDEBUGGER_READ_MEMORY)Irp->AssociatedIrp.SystemBuffer;
 
-            if (DebuggerCommandReadMemory(DebuggerReadMemRequest, DebuggerReadMemRequest, &ReturnSize) == TRUE)
+            if (DebuggerCommandReadMemory(DebuggerReadMemRequest,
+                                          ((CHAR *)DebuggerReadMemRequest) + SIZEOF_DEBUGGER_READ_MEMORY,
+                                          &ReturnSize) == TRUE)
             {
-                Status = STATUS_SUCCESS;
+                //
+                // Return the header a read bytes
+                //
+                Irp->IoStatus.Information = ReturnSize + SIZEOF_DEBUGGER_READ_MEMORY;
             }
             else
             {
-                Status = STATUS_UNSUCCESSFUL;
+                //
+                // Just return the header to the user-mode
+                //
+                Irp->IoStatus.Information = SIZEOF_DEBUGGER_READ_MEMORY;
             }
 
-            //
-            // Set the size
-            //
-            if (Status == STATUS_SUCCESS)
-            {
-                Irp->IoStatus.Information = ReturnSize;
+            Status = STATUS_SUCCESS;
 
-                //
-                // Avoid zeroing it
-                //
-                DoNotChangeInformation = TRUE;
-            }
+            //
+            // Avoid zeroing it
+            //
+            DoNotChangeInformation = TRUE;
 
             break;